% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-house-congress.R,
%   R/data-house-republic_act.R, R/data.R
\name{data_house_congress}
\alias{data_house_congress}
\alias{data_house_congress_load}
\alias{data_house_republic_act}
\alias{data}
\alias{data_bucket}
\alias{data_load}
\title{Legislative Data}
\usage{
data_house_congress()

data_house_congress_load()

data_house_republic_act()

data_bucket(path)

data_load()
}
\description{
Access the datasets stored in public DigitalOcean space. \verb{*_load()} functions
run the data extraction and load function into the space, for which you will
need access keys.
}
\section{Functions}{
\itemize{
\item \code{data_house_congress}: House of Representatives - Congress

\item \code{data_house_republic_act}: House of Representatives - Republic Acts

\item \code{data_bucket}: bucket storage

\item \code{data_load}: run the whole loading process
}}

