% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset.R
\name{dataset}
\alias{dataset}
\alias{dataset_write}
\alias{dataset_open}
\title{Read and Write Datasets to object storage}
\usage{
dataset_write(
  dataset,
  path,
  ...,
  policy = "public-read",
  creds = storage_credentials()
)

dataset_open(...)
}
\arguments{
\item{dataset, path, ...}{see \code{\link[arrow:write_dataset]{arrow::write_dataset}} or
\code{\link[arrow:open_dataset]{arrow::open_dataset}}}

\item{policy}{either \code{public-read} (default) or \code{private}.}

\item{creds}{storage credentials}
}
\description{
Write a dataset using the \code{\link[arrow:write_dataset]{arrow::write_dataset}} and then set the access
policy of the dataset (\code{public-read} by default). Read the dataset using
\code{\link[arrow:open_dataset]{arrow::open_dataset}}.
}
