% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset.R
\name{dataset_write}
\alias{dataset_write}
\title{Write a dataset to storage}
\usage{
dataset_write(
  dataset,
  path,
  ...,
  policy = "public-read",
  creds = storage_credentials()
)
}
\arguments{
\item{dataset, path, ...}{see \code{\link[arrow:write_dataset]{arrow::write_dataset}}}

\item{policy}{either \code{public-read} (default) or \code{private}.}

\item{creds}{storage credentials}
}
\description{
Write a dataset using the \code{\link[arrow:write_dataset]{arrow::write_dataset}} and then set the access
policy of the dataset (\code{public-read} by default).
}
