% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phlegis-package.R
\docType{package}
\name{phlegis-package}
\alias{phlegis}
\alias{phlegis-package}
\title{Philippines Legislative Analysis}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjpalanca/phlegis}
  \item Report bugs at \url{https://github.com/tjpalanca/phlegis/issues}
}

}
\author{
\strong{Maintainer}: Troy James Palanca \email{mail@tjpalanca.com} (\href{https://orcid.org/0000-0002-0610-9213}{ORCID})

}
\keyword{internal}
