% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-house-bill.R, R/data-house-congress.R,
%   R/data-house-republic_act.R, R/data-senate-bill.R, R/data-senate-congress.R,
%   R/data-senate-republic_act.R, R/data.R
\name{data_house_bill}
\alias{data_house_bill}
\alias{data_house_congress}
\alias{data_house_congress_pipeline}
\alias{data_house_republic_act}
\alias{data_house_republic_act_pipeline}
\alias{data_senate_bill}
\alias{data_senate_bill_pipeline}
\alias{data_senate_congress}
\alias{data_senate_congress_pipeline}
\alias{data_senate_republic_act}
\alias{data_senate_republic_act_pipeline}
\alias{data}
\alias{data_bucket}
\alias{data_pipeline}
\title{Legislative Data}
\usage{
data_house_bill()

data_house_congress()

data_house_congress_pipeline()

data_house_republic_act()

data_house_republic_act_pipeline(...)

data_senate_bill()

data_senate_bill_pipeline(...)

data_senate_congress()

data_senate_congress_pipeline()

data_senate_republic_act()

data_senate_republic_act_pipeline(...)

data_bucket(path)

data_pipeline()
}
\description{
Access the datasets stored in public DigitalOcean space.

The load process for these datasets is run by using the \verb{*_pipeline()}
functions. You will need the right access keys to be able to write to the
buckets.
}
\section{Functions}{
\itemize{
\item \code{data_house_bill}: House of Representatives - Bills and Resolutions

\item \code{data_house_congress}: House of Representatives - Congress

\item \code{data_house_republic_act}: House of Representatives - Republic Acts

\item \code{data_senate_bill}: Senate - Bills and Resolutions

\item \code{data_senate_congress}: Senate - Congress

\item \code{data_senate_republic_act}: Senate - Republic Acts

\item \code{data_bucket}: bucket storage

\item \code{data_pipeline}: run the whole data pipeline
}}

