% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-house-congress.R, R/data.R
\name{data_house_congress}
\alias{data_house_congress}
\alias{data_house_congress_load}
\alias{data}
\title{Legislative Data}
\usage{
data_house_congress()

data_house_congress_load()
}
\description{
Access the datasets stored in public DigitalOcean space. \verb{_load()} functions
run the data extraction and load function into the space, for which you will
need access keys.
}
\section{Functions}{
\itemize{
\item \code{data_house_congress}: House of Representatives - Congress
}}

